moved helper classes, global vars and exceptions to common.py
