+++ main.py
@@ -12,43 +12,7 @@ from datetime import timedelta
 
 # local imports
 from flowlauncher import FlowLauncher
-from exceptions import *
-
-formats = ["%d/%m/%Y", "%d/%-m/%Y", "%d-%m-%Y", "%d-%-m-%Y", "%d %m %Y", "%d %-m %Y", "%d.%m.%Y"]
-Week_Day={1:["Monday","Images/mon.png"],2:["Tuesday","Images/tue.png"],3:["Wednesday","Images/wed.png"],4:["Thursday","Images/thu.png"],5:["Friday","Images/fri.png"],6:["Saturday","Images/sat.png"],7:["Sunday","Images/sun.png"]}
-
-query_actions = {
-    "weekday": ["weekday", "wkd"],
-    "range": ["range", "rng"],
-    "unix_time": ["unixtime", "unx"],
-}
-
-class StrToDate:
-    """context manager to convert strings into datetime objects
-    using pre-set date formats. Will raise InvalidDateError if
-    one of the dates is not of a particular date.
-    """
-    def __init__(self, *datestrings):
-        self.datestrings = datestrings
-        print(self.datestrings)
-    
-    def __enter__(self):
-        dates = []
-        for datestr in self.datestrings:
-            dt = None
-            for format_str in formats:
-                try:
-                    dt = datetime.strptime(datestr, format_str)
-                except ValueError:
-                    continue
-            if dt is None:
-                raise InvalidDateError()
-            
-            dates.append(dt)
-        return dates
-    
-    def __exit__(self, exc_type, exc_val, traceback):
-        pass
+from common import *
 
 class DdFlow(FlowLauncher):
 
@@ -92,9 +56,8 @@ class DdFlow(FlowLauncher):
         except Exception as ex:
             print(ex)
             output.append({
-                "Title": "an unknown error had occured",
+                "Title": "",
                 "IcoPath": "Images/error.png"})
-                         
 
         return output
